fix(build): move test file out of pages directory
Next.js tries to bundle files in pages/ as routes. When vitest imports
are encountered, it fails because node:module can't be bundled for browser.

Move simulations-page.integration.test.tsx to src/features/simulations/
to avoid this issue.
