+++ langwatch/src/pages/[project]/simulations/index.tsx
@@ -8,7 +8,7 @@ import { PageLayout } from "~/components/ui/layouts/PageLayout";
 import { withPermissionGuard } from "~/components/WithPermissionGuard";
 import { useOrganizationTeamProject } from "~/hooks/useOrganizationTeamProject";
 import { api } from "~/utils/api";
-import { sortScenarioSets } from "./sort-scenario-sets";
+import { sortScenarioSets } from "~/features/simulations/sort-scenario-sets";
 
 function SimulationsPageContent() {
   const router = useRouter();
